Split interface and implementation in separate compilation units.
