+++ searchcore/src/vespa/searchcore/proton/persistenceengine/commit_and_wait_document_retriever.h
@@ -3,7 +3,7 @@
 #pragma once
 
 #include "i_document_retriever.h"
-#include <vespa/searchcore/proton/common/pendinglidtracker.h>
+#include <vespa/searchcore/proton/common/ipendinglidtracker.h>
 #include <vespa/searchcore/proton/common/icommitable.h>
 
 namespace proton {
@@ -17,7 +17,7 @@ class CommitAndWaitDocumentRetriever : public IDocumentRetriever
 {
     IDocumentRetriever::SP _retriever;
     ICommitable           &_commit;
-    ILidCommitState    &_uncommittedLidsTracker;
+    ILidCommitState       &_uncommittedLidsTracker;
     using Bucket = storage::spi::Bucket;
 public:
     CommitAndWaitDocumentRetriever(IDocumentRetriever::SP retriever, ICommitable &commit, ILidCommitState & unCommittedLidTracker);
